fix: make the rate limiter's bucket column a bigint
t.integer :bucket is 32-bit. Any rate_interval under 24 makes the bucket
width one second, so the bucket is the epoch second, which overflows int4
in January 2038. Changed both the migration template and the test harness
schema together, since the generator test checks them against each other.
Verified against real PostgreSQL and MySQL: reproduced PG::NumericValueOutOfRange
on the int4 column with a bucket value just past 2**31, then confirmed a
bigint column stores and reads it back correctly on both.

+++ test/support/active_record_database.rb
@@ -47,7 +47,7 @@ module ActiveRecordDatabase
   def self.define_rate_limits_table(connection)
     connection.create_table :translation_diff_rate_limits do |t|
       t.string :namespace, null: false, limit: 64
-      t.integer :bucket, null: false
+      t.bigint :bucket, null: false
       t.integer :characters, null: false, default: 0
     end
     connection.add_index :translation_diff_rate_limits, %i[namespace bucket],
